requirements.txt
Django==1.9.3
Naked==0.1.31
Pillow==3.1.1
PySide==1.2.4
PyYAML==3.11
image==1.4.4
pyasn1==0.1.9
pycrypto==2.6.1
requests==2.9.1
rsa==3.3
shellescape==3.4.1
stepic==bzr
termcolor==1.1.0
termrule==0.13
wsgiref==0.1.2
